requirements.txt
torch==1.10.0
numpy
imageio
gudhi==3.5.0
h5py
imagecodecs-lite==2020.1.31
Pillow==8.4.0
plyfile==0.7.4
PyMCubes==0.1.2
scikit-image==0.18.3
scikit-learn
scipy
SimpleITK==2.1.1
tensorboardX